--- budget-binder-common/src/androidMain/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="de.hsfl.budgetBinder.android"/>
