--- budget-binder-common/src/iosMain/kotlin/de.hsfl.budgetBinder.common/Platform.kt
@@ -1,7 +1,0 @@
-package de.hsfl.budgetBinder.common
-
-import platform.UIKit.UIDevice
-
-actual class Platform actual constructor() {
-    actual val platform: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
