+++ budget-binder-multiplatform-app/src/androidMain/kotlin/de/hsfl/budgetBinder/android/BudgetBinderApplication.kt
@@ -1,11 +1,6 @@
 package de.hsfl.budgetBinder.android
 
 import android.app.Application
-import de.hsfl.budgetBinder.common.HelloWorld
-import org.kodein.di.DI
-import org.kodein.di.DIAware
-import org.kodein.di.android.x.androidXModule
-import org.kodein.di.bindSingleton
 
 class BudgetBinderApplication: Application() {
