+++ budget-binder-multiplatform-app/src/androidMain/kotlin/de/hsfl/budgetBinder/model/Platform.kt
@@ -1,4 +1,4 @@
-package de.hsfl.budgetBinder.common
+package de.hsfl.budgetBinder.model
 
 actual class Platform actual constructor() {
     actual val platform: String = "Android ${android.os.Build.VERSION.SDK_INT}"
