+++ budget-binder-multiplatform-app/src/androidMain/res/drawable/hello_world.xml
