+++ budget-binder-multiplatform-app/src/commonMain/kotlin/de.hsfl.budgetBinder.model/Platform.kt
@@ -1,4 +1,4 @@
-package de.hsfl.budgetBinder.common
+package de.hsfl.budgetBinder.model
 
 // Platform anhängige implementierung
 expect class Platform() {
