+++ budget-binder-multiplatform-app/src/desktopMain/kotlin/de/hsfl/budgetBinder/model/Platform.kt
@@ -1,6 +1,5 @@
-package de.hsfl.budgetBinder.common
+package de.hsfl.budgetBinder.model
 
-// Platform anhängige implementierung
 actual class Platform actual constructor() {
     actual val platform: String = System.getProperty("os.name")
 }
