+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de.hsfl.budgetBinder.model/Platform.kt
@@ -1,5 +1,6 @@
-package de.hsfl.budgetBinder.common
+package de.hsfl.budgetBinder.model
 
+// Platform anhängige implementierung
 actual class Platform actual constructor() {
     actual val platform: String = "Kotlin JS"
 }
